Added docker compose and fixed persistence enable/disable scenarios
Fixed formatting

+++ src/iac_scan_runner/results_persistence.py
@@ -6,21 +6,32 @@ from datetime import datetime
 import os
 class ResultsPersistence:
     def __init__(self):
-    
         """
         Initialize new scan result database, collection and client
         """        
-        
+        self.connect_db()
+               
+    def connect_db(self):
+        """
+        Initialize new scan result database, collection and client
+        """
+                        
         try:
-            connection_string = os.environ['MONGO_STRING']  
-            print(connection_string)  
-            self.myclient = pymongo.MongoClient(connection_string)
-            self.mydb = self.myclient["scandb"]
-            self.mycol = self.mydb["results"]
-            self.connection_problem = False
+            connection_string = os.environ['MONGODB_CONNECTION_STRING']  
             
+            if(connection_string):
+                self.myclient = pymongo.MongoClient(connection_string)
+                self.mydb = self.myclient["scandb"]
+                self.mycol = self.mydb["results"]
+                self.connection_problem = False
+        
+        # TODO: Consider more specific exceptions     
         except Exception as e:
             print("Scan result persistence not available")
+            print(str(e))
+            self.myclient = None
+            self.mydb = None
+            self.mycol = None
             self.connection_problem = True
 
 
@@ -31,27 +42,38 @@ class ResultsPersistence:
     
         """Inserts new scan result into database
         :param result: Dictionary holding the scan summary
-        """         
-        self.mycol.insert_one(self.parse_json(result))
-
+        """     
+        if(self.connection_problem == True):
+            self.connect_db()
+        if(self.mycol != None):
+            self.mycol.insert_one(self.parse_json(result))
+         
     def show_result(self, uuid4: str) -> str:
     
         """Shows scan result with given id
         :param uuid4: Identifier of a scan result
         :return: String representing the scan result record
         """
-        myquery = { "uuid": uuid4 }
-        mydoc = self.mycol.find(myquery)
-        for x in mydoc:
-            return str(x)      
+        if(self.connection_problem==True):
+            self.connect_db()
+            
+        if(self.mycol != None):                
+            myquery = { "uuid": uuid4 }
+            mydoc = self.mycol.find(myquery)
+            for x in mydoc:
+                return str(x)      
 
     def delete_result(self, uuid4: str):
     
         """Deletes the scan result with given id from database
         :param uuid4: Identifier of a scan result which is about to be deleted
         """
-        myquery = { "uuid": uuid4 }
-        mydoc = self.mycol.delete_one(myquery)
+        if(self.connection_problem==True):
+            self.connect_db()
+            
+        if(self.mycol != None):          
+            myquery = { "uuid": uuid4 }
+            mydoc = self.mycol.delete_one(myquery)
 
 
     def show_all(self) -> str:
@@ -59,11 +81,15 @@ class ResultsPersistence:
         """Shows all the scan records from the database
         :return: String of all database records concatenated
         """
-        cursor = self.mycol.find({})
-        output = ""
-        for doc in cursor:
-            output = output + str(doc)
-        return output	                   
+        if(self.connection_problem==True):
+            self.connect_db()
+            
+        if(self.mycol != None):        
+            cursor = self.mycol.find({})
+            output = ""
+            for doc in cursor:
+                output = output + str(doc)
+            return output	                   
 
     def days_passed(self, time_stamp: str) -> int:
     
@@ -75,7 +101,8 @@ class ResultsPersistence:
         time2 = datetime.now() # current date and time
         delta = time2 - time1
         string_delta = str(delta)
-        if(string_delta.find("days")>-1):
+        
+        if(string_delta.find("days") > -1):
             days = string_delta.split(" ")
             days = days[0]
             return int(days)
@@ -89,22 +116,30 @@ class ResultsPersistence:
         :param uuid4: Identifier of a scan result
         :return: Integer denoting scan result age
         """
-        myquery = { "uuid": uuid4 }
-        mydoc = self.mycol.find(myquery)
-        for x in mydoc:
-            scan_ts = x["time"]
-
-        return self.days_passed(scan_ts)            
+        if(self.connection_problem == True):
+            self.connect_db()
+            
+        if(self.mycol != None):          
+            myquery = { "uuid": uuid4 }
+            mydoc = self.mycol.find(myquery)
+            for x in mydoc:
+                scan_ts = x["time"]
+                    
+            return self.days_passed(scan_ts)            
         
     def periodic_clean_job(self):
     
         """Calculates how long a scan result resides in database since its insertion
         :param uuid4: Identifier of a scan result
         """    
-        cursor = self.mycol.find({})
-        scan_ts = ""
-        for doc in cursor:
-            doc_uuid = doc["uuid"]                
-            age = self.result_age(doc_uuid)
-            if(age>14):
-                self.delete_result(doc_uuid)                                      
+        if(self.connection_problem == True):
+            self.connect_db()
+            
+        if(self.mycol != None):          
+            cursor = self.mycol.find({})
+            scan_ts = ""
+            for doc in cursor:
+                doc_uuid = doc["uuid"]                
+                age = self.result_age(doc_uuid)
+                if(age>14):
+                    self.delete_result(doc_uuid)
